Solved any tasks

+++ lib/controller/contact_controller.dart
@@ -1,10 +1,16 @@
 import 'dart:io';
 
+import 'package:permission_handler/permission_handler.dart';
+import 'package:qrscan/qrscan.dart' as scanner;
 import 'package:contactlist/database/database_fetch.dart';
 import 'package:contactlist/model/contact_model.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:image_picker/image_picker.dart';
+import 'package:geolocator/geolocator.dart';
+import 'package:url_launcher/url_launcher.dart';
+import 'package:flutter_share_me/flutter_share_me.dart';
+import 'package:simple_vcard_parser/simple_vcard_parser.dart';
 
 class ContactController extends GetxController {
   //ContactModel é uma lista observável, quando sofrer atualizações
@@ -12,13 +18,18 @@ class ContactController extends GetxController {
   var contactModel = []
       .obs; //.obs é responsável por informar ao Getx que a variável é observável e toda vez que ele sofrer alterações, deve ser comunicado os componentes que estão dentro do Obx()
   //declaração dos controladores de textos do formulário
-  TextEditingController nomeContactController = TextEditingController();
-  TextEditingController descricaoContactController = TextEditingController();
-  TextEditingController emailContactController = TextEditingController();
-  TextEditingController telefoneContactController = TextEditingController();
+  TextEditingController nomeContactController =
+      TextEditingController(text: 'Gabriel Rodrigo');
+  TextEditingController descricaoContactController =
+      TextEditingController(text: 'Mobile developer');
+  TextEditingController emailContactController =
+      TextEditingController(text: 'gabriel@actsistemas.com.br');
+  TextEditingController telefoneContactController =
+      TextEditingController(text: '4936214859');
+  TextEditingController siteContactController =
+      TextEditingController(text: 'https://actsistemas.com');
   TextEditingController latitudeContactController = TextEditingController();
   TextEditingController longitudeContactController = TextEditingController();
-  TextEditingController siteContactController = TextEditingController();
 
   GlobalKey<FormState> form = GlobalKey<FormState>();
 
@@ -57,6 +68,11 @@ class ContactController extends GetxController {
             nome: element['nome'],
             descricao: element['descricao'],
             foto: element['foto'],
+            telefone: element['telefone'],
+            email: element['email'],
+            latitude: element['latitude'],
+            longitude: element['longitude'],
+            site: element['site'],
           ),
         );
       });
@@ -65,37 +81,56 @@ class ContactController extends GetxController {
 
   void addData() async {
     if (form.currentState.validate()) {
+      if (image.value.isEmpty) {
+        Get.snackbar('Aviso', 'Tire uma foto para seu contato');
+        return;
+      }
+
+      Position position = await Geolocator.getCurrentPosition(
+          desiredAccuracy: LocationAccuracy.high);
       //grava na base de dados
-      var lastId = await DatabaseHelper.instance.insert(ContactModel(
+      latitudeContactController.text = position.latitude.toString();
+      longitudeContactController.text = position.longitude.toString();
+      saveData();
+      //fecha o formulário de cadastro
+      Get.back();
+    }
+  }
+
+  void saveData() async {
+    var lastId = await DatabaseHelper.instance.insert(ContactModel(
+      nome: nomeContactController.text,
+      descricao: descricaoContactController.text,
+      site: siteContactController.text,
+      telefone: telefoneContactController.text,
+      email: emailContactController.text,
+      foto: image.value,
+      latitude: latitudeContactController.text,
+      longitude: longitudeContactController.text,
+    ));
+    //insere os dados na lista atual que é exibida em tela
+    //evitando o reload da tabela
+    contactModel.insert(
+      0,
+      ContactModel(
+        id: lastId,
         nome: nomeContactController.text,
         descricao: descricaoContactController.text,
         site: siteContactController.text,
         telefone: telefoneContactController.text,
         email: emailContactController.text,
         foto: image.value,
-      ));
-      //insere os dados na lista atual que é exibida em tela
-      //evitando o reload da tabela
-      contactModel.insert(
-          0,
-          ContactModel(
-              id: lastId,
-              nome: nomeContactController.text,
-              descricao: descricaoContactController.text,
-              site: siteContactController.text,
-              telefone: telefoneContactController.text,
-              email: emailContactController.text,
-              foto: image.value));
-      //limpa os campos do formulário
-      nomeContactController.clear();
-      descricaoContactController.clear();
-      emailContactController.clear();
-      telefoneContactController.clear();
-      siteContactController.clear();
-      image.value = '';
-      //fecha o formulário de cadastro
-      Get.back();
-    }
+        latitude: latitudeContactController.text,
+        longitude: longitudeContactController.text,
+      ),
+    );
+    //limpa os campos do formulário
+    nomeContactController.clear();
+    descricaoContactController.clear();
+    emailContactController.clear();
+    telefoneContactController.clear();
+    siteContactController.clear();
+    image.value = '';
   }
 
   //recebe como parâmetro o id do registro no db e o caminho da foto, assim não precisa buscar o registro para resgatar o caminho da foto:
@@ -116,11 +151,37 @@ class ContactController extends GetxController {
         //remove os dados na lista atual que é exibida em tela
         //evitando o reload da tabela
         contactModel.removeWhere((element) => element.id == id);
-        await File(foto).delete(); //deleta o arquivo recebido por parâmetro
+        if (foto != "") {
+          //só remove um arquivo se existir imagem, pois pode ter sido importado contato via qrCode que não possui imagem.
+          await File(foto).delete(); //deleta o arquivo recebido por parâmetro
+        }
         Get.back(); //fecha o pop-up
       },
       onCancel: () => Get
           .back(), //adiciona automaticamente o botão Cancelar, e ao clicar fecha o pop-up
     );
+  }
+
+  void launchURL(url) async =>
+      await canLaunch(url) ? await launch(url) : throw 'Could not launch $url';
+
+  void shareOnWhatsapp(msg) async =>
+      await FlutterShareMe().shareToWhatsApp(msg: msg);
+
+  Future scan() async {
+    await Permission.camera.request();
+    String cameraScanResult = await scanner.scan();
+    VCard vc = VCard(cameraScanResult);
+    nomeContactController.text = vc.formattedName;
+    siteContactController.text = vc.typedURL[0][0];
+    telefoneContactController.text = vc.typedTelephone[0][0];
+    emailContactController.text = vc.typedEmail[0][0];
+
+    descricaoContactController.text = 'sem descricao';
+    image.value = '';
+    latitudeContactController.text = '';
+    longitudeContactController.text = '';
+
+    saveData();
   }
 }
